build: Use copy plugin to copy html & data (#60)
