+++ public/bookmarklets/index.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html lang="zh-Hant">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=0.9" />
+    <meta http-equiv="refresh" content="0; url=../#bookmarklets" />
+    <title>maimai Bookmarklets (Features & How to use)</title>
+  </head>
+  <body>
+    <p>
+      Redirecting...<br />
+      If page does not load in 3 seconds, please click
+      <a href="../#bookmarklets">this link</a>
+      to go to the new URL.
+    </p>
+    <p>
+      重新導向中。若三秒後沒有導向新網頁，請點
+      <a href="../#bookmarklets">此連結</a>
+      進入新網址。
+    </p>
+  </body>
+</html>
