+++ public/classic-layout/index.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width,initial-scale=1.0,user-scalable=yes,shrink-to-fit=no" >
+    <title>maimai classic score page</title>
+    <script src="adjustViewport.js"></script>
+  </head>
+  <body>
+    <div id="root"></div>
+    <script>var exports = {};</script>
+    <script crossorigin src="https://unpkg.com/react@16.13.1/umd/react.production.min.js"></script>
+    <script crossorigin src="https://unpkg.com/react-dom@16.13.1/umd/react-dom.production.min.js"></script>
+    <!-- <script src="../node_modules/react/umd/react.production.min.js"></script> -->
+    <!-- <script src="../node_modules/react-dom/umd/react-dom.production.min.js"></script> -->
+    <script src='./main.bundle.js'></script>
+  </body>
+</html>
