+++ public/data/chart-levels/version19.json
@@ -1,0 +1,42 @@
+{
+  "standard": {
+    "全世界共通リズム感テスト": [
+      -6,
+      -8,
+      -10,
+      12.0
+    ]
+  },
+  "dx": {
+    "ヒステリックナイトガール": [
+      -3,
+      -7,
+      -9.7,
+      12.8
+    ],
+    "INTERNET OVERDOSE": [
+      -4,
+      -7,
+      -10.7,
+      13.5
+    ],
+    "Knight Rider": [
+      -5,
+      -8,
+      11.2,
+      13.8
+    ],
+    "Let you DIVE!": [
+      -3,
+      -7.7,
+      -11.7,
+      12.8
+    ],
+    "Trrricksters!!": [
+      -6,
+      -8,
+      12.8,
+      14.5
+    ]
+  }
+}
