+++ public/data/song-info/intl.json
@@ -1,0 +1,26 @@
+{
+  "standard": {
+    "15": [
+      "全世界共通リズム感テスト"
+    ]
+  },
+  "dx": {
+    "15": [
+      "BREaK! BREaK! BREaK!"
+    ],
+    "16": [
+      "宿星審判"
+    ],
+    "19": [
+      "自傷無色",
+      "劣等上等",
+      "Let you DIVE!",
+      "Knight Rider",
+      "INTERNET OVERDOSE",
+      "Trrricksters!!",
+      "needLe",
+      "アイドル新鋭隊",
+      "セカイはまだ始まってすらいない"
+    ]
+  }
+}
