+++ public/rating-calculator/index.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=0.8" />
+    <title>maimai DX Rating Analyzer</title>
+    <script src="./adjustViewport.js"></script>
+  </head>
+  <body>
+    <div id="root"></div>
+    <script
+      crossorigin
+      src="https://cdnjs.cloudflare.com/ajax/libs/react/18.2.0/umd/react.production.min.js"
+    ></script>
+    <script
+      crossorigin
+      src="https://cdnjs.cloudflare.com/ajax/libs/react-dom/18.2.0/umd/react-dom.production.min.js"
+    ></script>
+    <!-- <script src="../node_modules/react/umd/react.development.js"></script> -->
+    <!-- <script src="../node_modules/react-dom/umd/react-dom.development.js"></script> -->
+    <script type="module" src="./main.bundle.js"></script>
+  </body>
+</html>
