+++ public/rating-visualizer/index.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=0.9" />
+    <title>maimai DX Rating Lookup Table & Visualization</title>
+  </head>
+  <body>
+    <div id="root"></div>
+    <script
+      crossorigin
+      src="https://cdnjs.cloudflare.com/ajax/libs/react/18.2.0/umd/react.production.min.js"
+    ></script>
+    <script
+      crossorigin
+      src="https://cdnjs.cloudflare.com/ajax/libs/react-dom/18.2.0/umd/react-dom.production.min.js"
+    ></script>
+    <!-- <script src="../node_modules/react/umd/react.development.js"></script> -->
+    <!-- <script src="../node_modules/react-dom/umd/react-dom.development.js"></script> -->
+    <script src="./main.bundle.js"></script>
+  </body>
+</html>
